refactor even more
