+++ src/modules/cats/catUtils.ts
@@ -1,6 +1,6 @@
 
-export const getFsImagePath = (name: string) => `src/public/cat/images/${name}`;
-export const getImagePath = (name: string) => `/cat/images/${name}`
+export const getFsImagePath = (name: string) => `public/img/cats/${name}`;
+export const getImagePath = (name: string) => `/img/cats/${name}`
 
 export const arrayRandom = <T extends any>(array: T[]) =>
   array[Math.floor(Math.random() * array.length)];
