+++ src/pages/api/cats/random/index.ts
@@ -1,7 +1,7 @@
-import { arrayRandom, getFsImagePath } from "../../../modules/cats/catUtils";
+import { arrayRandom, getFsImagePath } from "../../../../modules/cats/catUtils";
 
 import type { NextApiRequest, NextApiResponse } from "next";
-import { getImageNames } from "../../../serverUtils";
+import { getImageNames } from "../../../../serverUtils";
 import fs from "fs"
 
 export default async function handler(
