+++ src/serverUtils.ts
@@ -2,4 +2,4 @@ import path from "path";
 import { readdir } from "fs/promises";
 
 export const getImageNames = () =>
-  readdir(path.resolve(process.cwd(), "src/public/cat/images"));
+  readdir(path.resolve(process.cwd(), "public/img/cats"));
